Add files via upload
This project is a silly project where we have a button where we can add animals and once that button is pressed, the website undergoes a state change and once the website re-renders, a picture of a random animal is shown on the screen with a heart attached to the animal. If the user clicks on the animal, then the heart grows larger and larger. The user can keep adding as many animals as they like all with hearts attached to the animals.